Extract source inspection to internal package
